add htmlcontent getter and setter to say-view class
